requirements.txt
python-gnupg==0.5.1
pytest==7.4.0
mypy>=1.5.1
black>=23.7.0